# [IMP] base: make some tools lazy-loaded to speedup initialization