--- a/odoo/addons/base/models/res_partner.py
+++ b/odoo/addons/base/models/res_partner.py
@@ -1,15 +1,11 @@
-# -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 from __future__ import annotations
 
 import base64
 import collections
 import datetime
-import hashlib
 import pytz
-import re
 
-import requests
 from collections import defaultdict
 from random import randint
 from werkzeug import urls