Lazy import of some modules to speedup startup of base module.

- requests: 73ms (including 36ms for urllib3)
- zeep: 20ms + import of requests (or httpx)
- checking binary for wkhtmltopdf: 28ms (because it starts sub-processes)

---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr