@robodoo 2025-06-10
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/213532.png)](https://mergebot.odoo.com/odoo/odoo/pull/213532)